feat: separate types into groups
