+++ src/boolean.js
@@ -1,0 +1,11 @@
+const ensureBoolean = (value, defaultValue = false) => {
+  if (value === undefined || value === null) {
+    return Boolean(defaultValue);
+  }
+
+  return (typeof value === 'boolean') ? value : Boolean(value); // boolean coercible value
+};
+
+export {
+  ensureBoolean,
+};
