+++ src/object.js
@@ -1,0 +1,28 @@
+/**
+ * @param {any} obj The object to inspect.
+ * @returns {boolean} True if the argument appears to be a plain object.
+ */
+const isPlainObject = (obj) => {
+  if (typeof obj !== 'object' || obj === null) {
+    return false;
+  }
+
+  let proto = obj;
+  while (Object.getPrototypeOf(proto) !== null) {
+    proto = Object.getPrototypeOf(proto);
+  }
+
+  return Object.getPrototypeOf(obj) === proto;
+};
+
+const ensurePlainObject = (value, defaultValue = {}) => {
+  if (value === undefined || value === null) {
+    return ensurePlainObject(defaultValue);
+  }
+
+  return isPlainObject(value) ? value : ensurePlainObject(defaultValue);
+};
+
+export {
+  ensurePlainObject,
+};
