+++ src/string.js
@@ -1,0 +1,11 @@
+const ensureString = (value, defaultValue = '') => {
+  if (value === undefined || value === null) {
+    return String(defaultValue);
+  }
+
+  return (typeof value === 'string') ? value : String(value); // string coercible value
+};
+
+export {
+  ensureString,
+};
