+++ test/number.js
@@ -1,0 +1,56 @@
+import {
+  ensureNumber,
+  ensureFiniteNumber,
+} from '../src';
+
+describe('Number', () => {
+  test('ensureNumber', () => {
+    expect(ensureNumber()).toEqual(0);
+    expect(ensureNumber({})).toEqual(0);
+    expect(ensureNumber(true)).toEqual(1);
+    expect(ensureNumber(false)).toEqual(0);
+    expect(ensureNumber(0)).toEqual(0);
+    expect(ensureNumber(1)).toEqual(1);
+    expect(ensureNumber(2e+64)).toEqual(2e+64);
+    expect(ensureNumber(Infinity)).toEqual(Infinity);
+    expect(ensureNumber(-Infinity)).toEqual(-Infinity);
+    expect(ensureNumber(NaN)).toEqual(0);
+    expect(ensureNumber(undefined)).toEqual(0);
+    expect(ensureNumber(null)).toEqual(0);
+    expect(ensureNumber('0')).toEqual(0);
+    expect(ensureNumber('1')).toEqual(1);
+    expect(ensureNumber('')).toEqual(0);
+    expect(ensureNumber(' ')).toEqual(0);
+
+    // Returns the coerced default value.
+    expect(ensureNumber(null, '1')).toEqual(1);
+
+    // Returns the default value.
+    expect(ensureNumber(null, 1)).toEqual(1);
+  });
+
+  test('ensureFiniteNumber', () => {
+    expect(ensureFiniteNumber()).toEqual(0);
+    expect(ensureFiniteNumber({})).toEqual(0);
+    expect(ensureFiniteNumber(true)).toEqual(1);
+    expect(ensureFiniteNumber(false)).toEqual(0);
+    expect(ensureFiniteNumber(0)).toEqual(0);
+    expect(ensureFiniteNumber(1)).toEqual(1);
+    expect(ensureFiniteNumber(2e+64)).toEqual(2e+64);
+    expect(ensureFiniteNumber(Infinity)).toEqual(0);
+    expect(ensureFiniteNumber(-Infinity)).toEqual(0);
+    expect(ensureFiniteNumber(NaN)).toEqual(0);
+    expect(ensureFiniteNumber(undefined)).toEqual(0);
+    expect(ensureFiniteNumber(null)).toEqual(0);
+    expect(ensureFiniteNumber('0')).toEqual(0);
+    expect(ensureFiniteNumber('1')).toEqual(1);
+    expect(ensureFiniteNumber('')).toEqual(0);
+    expect(ensureFiniteNumber(' ')).toEqual(0);
+
+    // Returns the coerced default value.
+    expect(ensureFiniteNumber(null, NaN)).toEqual(0);
+
+    // Returns the default value.
+    expect(ensureFiniteNumber(null, 1)).toEqual(1);
+  });
+});
